refactor(ui): moved side nav compoent to layout folder
