--- src/app/side-nav-bar/side-nav-bar.component.html
@@ -1,10 +1,0 @@
-<section id="side-bar">
-  <div
-    id="menu-item"
-    *ngFor="let sideNavData of sideNavData"
-    (click)="sideMenuRoute(sideNavData.component)"
-  >
-    <img [src]="sideNavData.path" [alt]="sideNavData.name + ' icon'" />
-    <p id="menu-text">{{ sideNavData.name }}</p>
-  </div>
-</section>
